pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Chatbot Django + openrouter backend"
authors = [
    {name = "Jesus",email = "jesuga27@gmail.com"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "django-environ (>=0.13.0,<0.14.0)",
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "openai (>=2.32.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "channels (>=4.3.2,<5.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "drf-spectacular (>=0.29.0,<1.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.12.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pylint (>=3.0.0,<4.0.0)",
    "pylint-django (>=2.5.0,<3.0.0)",
    "autopep8 (>=2.0.0,<3.0.0)",
]
